requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-vlc==3.0.21203
websockets==15.0.1
